mining: drive the candidate lifecycle from MiningService
Move generation publication, long-poll waits, the bounded template
cache, and single-flight assembly into MiningService; node becomes the
capability provider:

- MiningService gains publish_generation(_from), notify_shutdown,
  live_generation_key, ensure_published, wait_for_generation_change,
  live_candidate / candidate_for_key / assemble_for_key, the
  Template-branch get_block_template, and mining_info. InFlight and the
  InFlightAssemblyGuard move with them; MempoolSequenceWake is
  implemented by the service.
- MempoolSnapshotSource::mining_snapshot_at captures the snapshot under
  the same pool read lock that checks the expected sequence, preserving
  the original generation-key coherence window.
- MiningCoordinator composes the service: it builds the
  AppliedTip/Mempool/ChainContext adapters, keeps proposal validation
  and solved-block submission, and forwards publish_generation_from /
  notify_shutdown / Template / mining_info assembly. The node-owned
  facts (applied-tree hashrate walk, node metrics warnings) stay in the
  facade and feed service.mining_info.
- long_poll.rs dissolves into the service; MAX_BLOCK_WEIGHT/SIZE move
  to crates/mining; the node clock read switches to
  bitcoin_rs_chain::current_unix_seconds (same u32 clamping as the
  coordinator's inlined copy, which is deleted).

Verified: cargo check -p bitcoin-rs-mining -p bitcoin-rs-node
-p bitcoin-rs-rpc; cargo test -p bitcoin-rs-mining;
cargo test -p bitcoin-rs-node --lib mining (25 passed);
cargo test -p bitcoin-rs-node --test mining (49 passed).
